requirements.txt
numpy==2.3.1
pygame-ce==2.5.5
pygame_gui==0.6.14
python-i18n==0.3.9
watchdog==6.0.0
